LayoutTests/http/wpt/crypto/historical.html

<!DOCTYPE html>
<html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
// localhost is secure by default.
internals.markContextAsInsecure();

test(() => {
  assert_equals(self.crypto.subtle, undefined);
  assert_false("subtle" in self.crypto);
}, "Non-secure context window does not have access to crypto.subtle");

test(() => {
  assert_equals(self.SubtleCrypto, undefined);
  assert_false("SubtleCrypto" in self);
}, "Non-secure context window does not have access to SubtleCrypto")

test(() => {
  assert_equals(self.CryptoKey, undefined);
  assert_false("CryptoKey" in self);
}, "Non-secure context window does not have access to CryptoKey")
</script>
</body>
</html>
